--- a/src/main/php/util/data/Marshalling.class.php
+++ b/src/main/php/util/data/Marshalling.class.php
@@ -1,26 +1,18 @@
 <?php namespace util\data;
 
-use lang\ArrayType;
-use lang\Enum;
-use lang\MapType;
-use lang\Type;
-use lang\XPClass;
-use util\Bytes;
-use util\Currency;
-use util\Date;
-use util\Money;
-use util\XPIterator;
+use lang\{ArrayType, Enum, MapType, Reflection, Type, XPClass};
+use util\{Bytes, Currency, Date, Money, XPIterator};
 
 /**
  * Takes care of converting objects from and to maps
  *
- * @test  xp://util.data.unittest.MarshallingTest
- * @test  xp://util.data.unittest.PrimitivesTest
- * @test  xp://util.data.unittest.BytesTest
- * @test  xp://util.data.unittest.DatesTest
- * @test  xp://util.data.unittest.EnumsTest
- * @test  xp://util.data.unittest.MoneyTest
- * @test  xp://util.data.unittest.ObjectsTest
+ * @test  util.data.unittest.MarshallingTest
+ * @test  util.data.unittest.PrimitivesTest
+ * @test  util.data.unittest.BytesTest
+ * @test  util.data.unittest.DatesTest
+ * @test  util.data.unittest.EnumsTest
+ * @test  util.data.unittest.MoneyTest
+ * @test  util.data.unittest.ObjectsTest
  */
 class Marshalling {
 
@@ -35,18 +27,6 @@ class Marshalling {
     foreach ($in as $key => $value) {
       yield $key => $this->unmarshal($value, $type);
     }
-  }
-
-  /**
-   * Check whether a type's constructor accepts a given value
-   *
-   * @param  lang.Type $type
-   * @param  var $value
-   * @return bool
-   */
-  private function constructorAccepts($type, $value) {
-    $params= $type->getConstructor()->getParameters();
-    return 1 === sizeof($params) && $params[0]->getType()->isInstance($value);
   }
 
   /**
@@ -77,30 +57,33 @@ class Marshalling {
         return new Iteration($value);
       } else if ($t->isInterface()) {
         return $t->cast($value);
-      } else if ($t->hasConstructor() && $this->constructorAccepts($t, $value)) {
-        return $t->newInstance($value);
       }
 
-      $r= $t->reflect()->newInstanceWithoutConstructor();
+      $reflect= Reflection::type($t);
+
+      // If a single-argument constructor accepting the value exists, invoke it.
+      if (
+        ($constructor= $reflect->constructor()) &&
+        (1 === $constructor->parameters()->size()) &&
+        $constructor->parameter(0)->constraint()->type()->isInstance($value)
+      ) {
+        return $constructor->newInstance([$value]);
+      }
+
+      $r= $reflect->initializer(null)->newInstance();
       if (method_exists($r, '__unserialize')) {
         $r->__unserialize($value);
         return $r;
       }
 
-      foreach ($t->getFields() as $field) {
-        $m= $field->getModifiers();
-        if ($m & MODIFIER_STATIC) continue;
-
-        $n= $field->getName();
-        if (!isset($value[$n])) continue;
-
-        if ($m & MODIFIER_PUBLIC) {
-          $field->set($r, $this->unmarshal($value[$n], $field->getType()));
-        } else if ($t->hasMethod($set= 'set'.ucfirst($n))) {
-          $method= $t->getMethod($set);
-          $method->invoke($r, [$this->unmarshal($value[$n], $method->getParameter(0)->getType())]);
+      foreach ($reflect->properties() as $name => $p) {
+        $modifiers= $p->modifiers();
+        if ($modifiers->isStatic() || !isset($value[$name])) {
+          continue;
+        } else if ($m= $reflect->method('set'.ucfirst($name))) {
+          $m->invoke($r, [$this->unmarshal($value[$name], $m->parameter(0)->constraint()->type())]);
         } else {
-          $field->setAccessible(true)->set($r, $this->unmarshal($value[$n], $field->getType()));
+          $p->set($r, $this->unmarshal($value[$name], $p->constraint()->type()), $reflect);
         }
       }
       return $r;
@@ -175,22 +158,16 @@ class Marshalling {
       if (method_exists($value, '__toString')) return $value->__toString();
 
       $r= [];
-      $type= typeof($value);
-      foreach ($type->getFields() as $field) {
-        $m= $field->getModifiers();
-        if ($m & MODIFIER_STATIC) continue;
-
-        $n= $field->getName();
-        if ($m & MODIFIER_PUBLIC) {
-          $v= $field->get($value);
-        } else if ($type->hasMethod($n)) {
-          $v= $type->getMethod($n)->invoke($value, []);
-        } else if ($type->hasMethod($get= 'get'.ucfirst($n))) {
-          $v= $type->getMethod($get)->invoke($value, []);
+      $reflect= Reflection::type($value);
+      foreach ($reflect->properties() as $name => $p) {
+        $modifiers= $p->modifiers();
+        if ($modifiers->isStatic()) {
+          continue;
+        } else if ($m= $reflect->method($name) ?? $reflect->method('get'.ucfirst($name))) {
+          $r[$name]= $this->marshal($m->invoke($value, []));
         } else {
-          $v= $field->setAccessible(true)->get($value);
+          $r[$name]= $this->marshal($p->get($value, $reflect));
         }
-        $r[$n]= $this->marshal($v);
       }
       return $r;
     } else if (is_array($value)) {