See https://github.com/xp-framework/rfc/issues/338

## Performance

Before:
```
Test cases:  72 succeeded
Memory used: 2694.68 kB (2748.81 kB peak)
Time taken:  0.012 seconds (0.058 seconds overall)
```

After:
```
Test cases:  72 succeeded
Memory used: 2438.81 kB (2492.94 kB peak)
Time taken:  0.011 seconds (0.053 seconds overall)
```